src/ai_solver_experiment/requirements.txt
anthropic>=0.39.0
openai>=1.50.0
google-generativeai>=0.8.0
